notebook for scaleout plots
