+++ .gitignore
@@ -46,3 +46,6 @@ evaluation/plot/etc/
 evaluation/plot/lib/
 evaluation/plot/pyvenv.cfg
 evaluation/plot/share/
+
+__pycache__
+evaluation/plot/*.pdf
